0 errors, 0 warnings in vc141. CMake: release copies correct dlls. Updated to glew 2.1.0. Changed CRT link to static.

+++ sources/effects/efxGlow2.c
@@ -21,7 +21,7 @@ void preload_efxGlow2 () {
 
 void load_efxGlow2 () {
 	char *equation;
-	int equationSize;
+	size_t equationSize;
 	
 	// script validation
 	if (mySection->stringNum != 5) {
